Fixed importing and exporting issue
Catch date exporting was only accurate to the minute (to keep
compatibility with iOS).  It has been updated to be accurate to a
thousandth of a second, but will still parse the old format if needed.
If there is a duplicate date as a result of the inaccurate date format,
a random 60000 ms is added to the date, then another random 60000 ms is
removed from the date.  After the alteration, the chances of duplicate
dates are basically zero.